cleaning up ProfilePage + adding Delete Account Modal
